refactor: move posts dir maker to new file
